Moved'Demo User Login' button within Login/Signup forms
