+++ client/src/pages/Login/LoginForm/useStyles.ts
@@ -35,12 +35,12 @@ const useStyles = makeStyles((theme) => ({
     padding: '0 10px',
   },
   submit: {
-    margin: theme.spacing(3, 2, 2),
+    margin: theme.spacing(2),
     padding: 10,
     width: 200,
     height: 45,
     borderRadius: theme.shape.borderRadius,
-    marginTop: 49,
+    marginTop: 10,
     fontSize: 14,
     backgroundColor: theme.palette.primary.main,
     color: '#ffffff',
@@ -49,6 +49,10 @@ const useStyles = makeStyles((theme) => ({
       width: '100%',
       margin: '24px auto 0',
     },
+  },
+  demoLoginBtn: {
+    backgroundColor: '#ffffff',
+    color: theme.palette.primary.main,
   },
   underline: {
     '&&:after': {
